Consolidate KdcOptionFlags and TicketFlags
