+++ lib/msf/core/exploit/remote/kerberos/service_authenticator/base.rb
@@ -282,10 +282,10 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
 
     ticket_options = Rex::Proto::Kerberos::Model::KdcOptionFlags.from_flags(
       [
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::FORWARDABLE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::CANONICALIZE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE_OK
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::FORWARDABLE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::CANONICALIZE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE_OK
       ]
     )
 
@@ -330,7 +330,7 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
     when Delegation::ALWAYS
       do_delegation = true
     when Delegation::WHEN_UNCONSTRAINED
-      do_delegation = credential.ticket_flags.include?(Rex::Proto::Kerberos::Model::KdcOptionFlag::OK_AS_DELEGATE)
+      do_delegation = credential.ticket_flags.include?(Rex::Proto::Kerberos::Model::KdcOptionFlags::OK_AS_DELEGATE)
     end
 
     if do_delegation
@@ -352,10 +352,10 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
       client_name: username,
       options: Rex::Proto::Kerberos::Model::KdcOptionFlags.from_flags(
         [
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::FORWARDABLE,
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE,
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::CANONICALIZE,
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE_OK
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::FORWARDABLE,
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE,
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::CANONICALIZE,
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE_OK
         ]
       ),
       sequence_number: sequence_number
@@ -404,7 +404,7 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
     when Delegation::NEVER
       do_delegation = false
     when Delegation::WHEN_UNCONSTRAINED
-      do_delegation = tgs_auth.flags.include?(Rex::Proto::Kerberos::Model::KdcOptionFlag::OK_AS_DELEGATE)
+      do_delegation = tgs_auth.flags.include?(Rex::Proto::Kerberos::Model::KdcOptionFlags::OK_AS_DELEGATE)
     end
 
     if do_delegation
@@ -441,10 +441,10 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
       client_name: client_name,
       options: Rex::Proto::Kerberos::Model::KdcOptionFlags.from_flags(
         [
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::FORWARDABLE,
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE,
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::CANONICALIZE,
-          Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE_OK
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::FORWARDABLE,
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE,
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::CANONICALIZE,
+          Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE_OK
         ]
       ),
       sequence_number: sequence_number,
@@ -515,10 +515,10 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
   def request_delegation_ticket(session_key, tgt_ticket, realm, client_name, tgt_etype, expiry_time, now)
     ticket_options = Rex::Proto::Kerberos::Model::KdcOptionFlags.from_flags(
       [
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::FORWARDABLE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::FORWARDED,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::CANONICALIZE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::FORWARDABLE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::FORWARDED,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::CANONICALIZE,
       ]
     )
 
@@ -586,9 +586,9 @@ class Msf::Exploit::Remote::Kerberos::ServiceAuthenticator::Base
   def request_service_ticket(session_key, tgt_ticket, realm, client_name, tgt_etype, expiry_time, now, sname)
     ticket_options = Rex::Proto::Kerberos::Model::KdcOptionFlags.from_flags(
       [
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::FORWARDABLE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::RENEWABLE,
-        Rex::Proto::Kerberos::Model::KdcOptionFlag::CANONICALIZE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::FORWARDABLE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::RENEWABLE,
+        Rex::Proto::Kerberos::Model::KdcOptionFlags::CANONICALIZE,
       ]
     )
