+++ lib/msf/core/exploit/remote/kerberos/ticket.rb
@@ -88,15 +88,15 @@ module Msf
 
           def silver_ticket_flags
             [
-              Rex::Proto::Kerberos::Model::TicketFlag::FORWARDABLE,
-              Rex::Proto::Kerberos::Model::TicketFlag::PROXIABLE,
-              Rex::Proto::Kerberos::Model::TicketFlag::RENEWABLE,
-              Rex::Proto::Kerberos::Model::TicketFlag::PRE_AUTHENT
+              Rex::Proto::Kerberos::Model::TicketFlags::FORWARDABLE,
+              Rex::Proto::Kerberos::Model::TicketFlags::PROXIABLE,
+              Rex::Proto::Kerberos::Model::TicketFlags::RENEWABLE,
+              Rex::Proto::Kerberos::Model::TicketFlags::PRE_AUTHENT
             ]
           end
 
           def golden_ticket_flags
-            silver_ticket_flags << Rex::Proto::Kerberos::Model::TicketFlag::INITIAL
+            silver_ticket_flags << Rex::Proto::Kerberos::Model::TicketFlags::INITIAL
           end
 
           def generate_ccache(ticket, opts:)
