+++ spec/lib/msf/core/exploit/remote/kerberos/ticket_spec.rb
@@ -62,10 +62,10 @@ RSpec.describe Msf::Exploit::Remote::Kerberos::Ticket do
     let(:ticket_key) { [nt_hash].pack('H*') }
     let(:flags) do
       [
-        Rex::Proto::Kerberos::Model::TicketFlag::FORWARDABLE,
-        Rex::Proto::Kerberos::Model::TicketFlag::PROXIABLE,
-        Rex::Proto::Kerberos::Model::TicketFlag::RENEWABLE,
-        Rex::Proto::Kerberos::Model::TicketFlag::PRE_AUTHENT
+        Rex::Proto::Kerberos::Model::TicketFlags::FORWARDABLE,
+        Rex::Proto::Kerberos::Model::TicketFlags::PROXIABLE,
+        Rex::Proto::Kerberos::Model::TicketFlags::RENEWABLE,
+        Rex::Proto::Kerberos::Model::TicketFlags::PRE_AUTHENT
       ]
     end
 
@@ -113,10 +113,10 @@ RSpec.describe Msf::Exploit::Remote::Kerberos::Ticket do
         describe 'ticket enc part' do
           let(:expected_flags) do
             [
-              Rex::Proto::Kerberos::Model::TicketFlag::FORWARDABLE,
-              Rex::Proto::Kerberos::Model::TicketFlag::PROXIABLE,
-              Rex::Proto::Kerberos::Model::TicketFlag::RENEWABLE,
-              Rex::Proto::Kerberos::Model::TicketFlag::PRE_AUTHENT
+              Rex::Proto::Kerberos::Model::TicketFlags::FORWARDABLE,
+              Rex::Proto::Kerberos::Model::TicketFlags::PROXIABLE,
+              Rex::Proto::Kerberos::Model::TicketFlags::RENEWABLE,
+              Rex::Proto::Kerberos::Model::TicketFlags::PRE_AUTHENT
             ]
           end
